requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
cx_Oracle==8.3.0
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozendict==2.4.4
html5lib==1.1
hvac==2.3.0
idna==3.7
lxml==5.2.2
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==2.0.0
packaging==24.1
pandas==2.2.2
peewee==3.17.5
pillow==10.4.0
platformdirs==4.2.2
plotly==5.22.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tenacity==8.5.0
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.40
